# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Chinook API",
        "contact": {
            "name": "Testsmith",
            "email": "info@testsmith.io"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Local environment"
        }
    ],
    "paths": {
...
        "/mediatypes": {
            "get": {
                "tags": [
                    "Mediatype"
                ],
                "summary": "Retrieve all mediatypes",
                "operationId": "get-all-mediatypes",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page, default=1",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All mediatypes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MediatypeResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Mediatype"
                ],
                "summary": "New mediatype",
                "operationId": "store-mediatype",
                "requestBody": {
                    "description": "Mediatype object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MediatypeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An Album",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MediatypeResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/mediatypes/{mediatypeId}": {
            "get": {
                "tags": [
                    "Mediatype"
                ],
                "summary": "Retrieve a specific mediatype",
                "operationId": "get-mediatype",
                "parameters": [
                    {
                        "name": "mediatypeId",
                        "in": "path",
                        "description": "The mediatypeId parameter in path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An Mediatype",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MediatypeResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            },
            "put": {
                "tags": [
                    "Mediatype"
                ],
                "summary": "Update mediatype",
                "operationId": "update-mediatype",
                "parameters": [
                    {
                        "name": "mediatypeId",
                        "in": "path",
                        "description": "The mediatypeId parameter in path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Mediatype object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MediatypeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An Mediatype",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MediatypeResponse"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Mediatype"
                ],
                "summary": "Remove mediatype",
                "operationId": "destroy-mediatype",
                "parameters": [
                    {
                        "name": "mediatypeId",
                        "in": "path",
                        "description": "The mediatypeId parameter in path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
        "/mediatypes/search": {
            "get": {
                "tags": [
                    "Mediatype"
                ],
                "summary": "Retrieve specific mediatypes matching the search query",
                "operationId": "search-mediatype",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "The query parameter in path",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All matching mediatypes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MediatypeResponse"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MediatypeRequest": {
                "title": "MediatypeRequest",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "description": "Class Mediatype",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "MediatypeResponse": {
                "title": "MediatypeResponse",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}